HBP-157 Added kubecontroller; Removed bouncer in front flag
