+++ framework_tvb/tvb/interfaces/web/operation_openshift_assigner.py
@@ -60,7 +60,8 @@ if __name__ == '__main__':
                     for index, operation in enumerate(operations):
                         pod_ip = pods[index % pods_no]['ip']
                         log.info("Notify pod: {}".format(pod_ip))
-                        url_pattern = "http://{}:8080/user/start_operation_pod/{}"
-                        executor.submit(requests.get, url=url_pattern.format(pod_ip, operation.id), headers=auth_header)
+                        url_pattern = "http://{}:{}/kube/start_operation_pod/{}"
+                        executor.submit(requests.get, url=url_pattern.format(pod_ip, TvbProfile.current.web.SERVER_PORT,
+                                                                             operation.id), headers=auth_header)
     else:
         log.info("Openshift deploy is not enabled.")
