+++ framework_tvb/tvb/interfaces/web/run.py
@@ -37,6 +37,7 @@ import time
 
 from tvb.storage.storage_interface import StorageInterface
 from tvb.core.services.backend_clients.standalone_client import StandAloneClient
+from tvb.interfaces.web.controllers.kube_controller import KubeController
 
 STARTUP_TIC = time.time()
 
@@ -153,6 +154,7 @@ def init_cherrypy(arguments=None):
     cherrypy.tree.mount(LocalConnectivityController(), "/spatial/localconnectivity/", config=CONFIGUER)
     cherrypy.tree.mount(NoiseConfigurationController(), "/burst/noise/", config=CONFIGUER)
     cherrypy.tree.mount(HPCController(), "/hpc/", config=CONFIGUER)
+    cherrypy.tree.mount(KubeController(), "/kube/", config=CONFIGUER)
 
     cherrypy.config.update(CONFIGUER)
