+++ scientific_library/tvb/basic/config/stored.py
@@ -60,7 +60,6 @@ KEY_SELECTED_DB = 'SELECTED_DB'
 KEY_DB_URL = 'URL_VALUE'
 KEY_MAX_CONNECTIONS = 'MAX_CONNECTIONS'
 KEY_MAX_ASYNC_CONNECTIONS = 'MAX_ASYNC_CONNECTIONS'
-KEY_BOUNCER = 'BOUNCER_IN_FRONT'
 KEY_URL_VERSION = 'URL_TVB_VERSION'
 KEY_CLUSTER = 'DEPLOY_CLUSTER'
 KEY_CLUSTER_SCHEDULER = 'CLUSTER_SCHEDULER'
