+++ tvb_build/docker/.tvb.configuration
@@ -27,7 +27,10 @@ CRYPT_DATADIR=/home/tvb_user/.data
 CRYPT_PASSDIR=/home/tvb_user/.pass
 ENCRYPT_STORAGE=False
 DECRYPT_PATH=/home/tvb_user/DECRYPT
-KEY_MAX_CONNECTIONS = 'MAX_CONNECTIONS'
-KEY_MAX_ASYNC_CONNECTIONS = 'MAX_ASYNC_CONNECTIONS'
-KEY_BOUNCER = 'BOUNCER_IN_FRONT'
+MAX_CONNECTIONS=20
+MAX_ASYNC_CONNECTION=2
+OPENSHIFT_DEPLOY=False
+OPENSHIFT_NAMESPACE=tvb-crypt
+OPENSHIFT_APPLICATION=thevirtualbrain
+PROCESSING_OPERATIONS_APPLICATION=process-operations
 FILE_STORAGE=h5
